feat: 게시물 리스트 파일 구조 변경 (#328)
