+++ frontend/src/components/Apply/Filter/DetailFilter/index.jsx
