+++ frontend/src/components/Apply/Filter/DetailFilter/range.css
